Testing on search
I did some testing on the search algorithm to try and familiarize myself
with it. Added search for Miami name as well as searching through
cultural uses to try and see if it would catch names of the sources
archival references. Also edited show.blade.php to not show the
researchers note (one of the items they said we could ditch)