Updated the advanced search page and css file
Updated the advanced search page to have a dropdown menu for each
category that the user can pick which item s/he wants to search on.
Should probably try to include a generic 'choose a _______' as the
default incase the user doesn't want to use a specific field. Search
capabilities don't work at all, but baby steps right?